Create project structure using cylc namespace
