+++ cylc/uiserver/data_mgr.py
